Start handling Heroku cookies more seriously
We've noticed a problem across Heroku properties that seems to result
from a desynchronization between the expiry of the standard Identity
cookie and Heroku cookies (`heroku_session` + `heroku_session_nonce`).
This is because although the standard cookie is refreshed on every
request thanks to middleware, the Heroku cookie is only set a single
time. This moves Heroku cookie logic to another piece of custom
middleware that behaves like standard cookie middleware so that the
application doesn't have to deal with cookie logic.