OAuth Callback Assertions
OAuth callback assertions have a number of hard-coded `code` values
and URLs. This adds a custom assertion to pull up the hard-coded
test values into a constant and check the constants when the
`assert_response_redirects_with_oauth_callback` assertion is called.
An optional state value can be passed to the custom assertion.